# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The public api service is the service that consumes all incoming requests.",
    "version": "1.0.0",
    "title": "2ndLock public api service",
    "termsOfService": "http://2ndLock.com",
    "contact": {
      "email": "jens@symlink.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/password/forgot": {
      "post": {
        "tags": ["password"],
        "summary": "send forgot password request",
        "description": "send forgot password request",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ForgotPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/password/update": {
      "post": {
        "tags": ["password"],
        "summary": "send new password request",
        "description": "send new password request",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ForgotPassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "UpdatePassword": {
      "type": "object",
      "properties": {
        "otp": {
          "type": "string"
        },
        "forgotPasswordId": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
...
  }
}